Enhance provider classes and remove duplicates
It was obvious that there are many duplicates between Metamask and
WalletConnect sub-classes, this has been resolved by creating an
abstract class EthereumProvider that contains all shared code between
the two providers sub-classes, then instead of implementing interface
IEthereumProvider in sub-classes, it's imeplemented on parent class,
and then parent is extended for different wallet. Because requestSigner
method is unique (at least till the moment) for MetaMask, it has been
separated into ISignerinterface that only MetaMaskProvider sub-class
implements it.
