+++ frontend/src/services/web3-provider/index.ts
@@ -1,3 +1,4 @@
+export * from './ethereum-provider';
 export * from './metamask-provider';
 export * from './types';
 export * from './walletconnect-provider';
