+++ frontend/tests/utils/mocks/ethereum-provider.ts
@@ -2,9 +2,9 @@ import { JsonRpcSigner } from '@ethersproject/providers';
 import type { Ref, ShallowRef } from 'vue';
 import { ref, shallowRef } from 'vue';
 
-import { EthereumProvider } from '@/services/web3-provider';
+import { IEthereumProvider, ISigner } from '@/services/web3-provider';
 
-export class MockedEthereumProvider implements EthereumProvider {
+export abstract class MockedEthereumProvider implements IEthereumProvider {
   readonly signer: ShallowRef<JsonRpcSigner | undefined>;
   readonly signerAddress: ShallowRef<string | undefined>;
   readonly chainId: Ref<number>;
@@ -16,11 +16,24 @@ export class MockedEthereumProvider implements EthereumProvider {
   }
 
   init = vi.fn();
-  requestSigner = vi.fn();
   getLatestBlock = vi.fn();
   connectContract = vi.fn();
   switchChain = vi.fn();
   addChain = vi.fn();
   getChainId = vi.fn(async () => this.chainId.value);
   addToken = vi.fn();
+}
+
+export class MockedMetMaskProvider extends MockedEthereumProvider implements ISigner {
+  constructor(options?: { chainId?: number; signer?: JsonRpcSigner; signerAddress?: string }) {
+    super(options);
+  }
+
+  requestSigner = vi.fn();
+}
+
+export class MockedWalletConnectProvider extends MockedEthereumProvider {
+  constructor(options?: { chainId?: number; signer?: JsonRpcSigner; signerAddress?: string }) {
+    super(options);
+  }
 }
